requirements.txt
requests==2.31.0
python-dotenv==1.0.0
supabase==2.4.1
google-generativeai==0.5.4
pandas==2.2.2
beautifulsoup4==4.12.3
lxml==5.2.2